requirements.txt
django
pymysql
mysqlclient
djangorestframework
djangorestframework-jwt
django-cors-headers
